+++ NEWS.md
@@ -1,7 +1,12 @@
-# ggghost 0.2.0
+# CHANGES in ggghost 0.2.0
 
-* 
 
-# ggghost 0.1.0 
+
+* Added a warning to `recover_data` when data object exists in the calling frame
+but has changed since being captured (closes #4).
+
+* Moved README supplementary objects out of package (closes #2).
+
+# ggghost 0.1.0
 
 * Initial CRAN submission. No ERRORs, WARNINGs, or NOTEs.